migration-simulation/pyproject.toml
[project]
name = "migration-simulation"
version = "0.1.0"
description = "E2E verification tests for the NoUI migration engine"
requires-python = ">=3.11"

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pyyaml>=6.0",
    "requests>=2.28",
    "psycopg2-binary>=2.9",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
